LayoutTests/webgpu/render-pipelines.html

<!DOCTYPE html>
<html>
<script src="js/webgpu-functions.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
const shaderCode = `
#include <metal_stdlib>
    
using namespace metal;

struct Vertex
{
    float4 position [[position]];
};

vertex Vertex vertex_main(uint vid [[vertex_id]])
{
    Vertex v;
    switch (vid) {
    case 0:
        v.position = float4(-1, 1, 0, 1);
        break;
    case 1:
        v.position = float4(-1, -1, 0, 1);
        break;
    case 2:
        v.position = float4(1, 1, 0, 1);
        break;
    default:
        v.position = float4(1, -1, 0, 1);
    }
    return v;
}

fragment float4 fragment_main(Vertex vertexIn [[stage_in]])
{
    return float4(0.0, 1.0, 0.0, 1.0);
}
`

promise_test(async () => {
    const device = await getBasicDevice();
    const shaderModule = device.createShaderModule({ code: shaderCode });

    const layoutBinding = { binding: 0, visibility: WebGPUShaderStageBit.VERTEX, type: "storageBuffer" };
    const bindGroupLayout = device.createBindGroupLayout({ bindings: [layoutBinding] });
    const pipelineLayout = device.createPipelineLayout({ bindGroupLayouts: [bindGroupLayout] });

    const pipeline = createBasicPipeline(shaderModule, device, pipelineLayout);
    assert_true(pipeline instanceof WebGPURenderPipeline, "Successfully created WebGPURenderPipeline");
}, "Create basic WebGPURenderPipeline");

function checkBadRenderPipeline(device, descriptor, testSubjectName) {
    assert_true(!device.createRenderPipeline(descriptor), `WebGPURenderPipeline with invalid ${testSubjectName} was not created.`);
}

promise_test(async() => {
    const device = await getBasicDevice();
    const shaderModule = device.createShaderModule({ code: shaderCode });
    checkBadRenderPipeline(device, {}, "WebGPURenderPipelineDescriptor");

    const noModuleVertexDescriptor = {
        entryPoint: "vertex_main"
    }

    const fragmentStageDescriptor = {
        module: shaderModule,
        entryPoint: "fragment_main"
    };

    const noModulePipelineDescriptor = {
        vertexStage: noModuleVertexDescriptor, 
        fragmentStage: fragmentStageDescriptor,
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(device, noModulePipelineDescriptor, "shader module");

    // A (Metal) renderpipeline must have a vertex function.
    const noVertexPipelineDescriptor = {
        fragmentStage: fragmentStageDescriptor,
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(device, noVertexPipelineDescriptor, "vertex shader stage")

    const badEntryPointDescriptor = { 
        module: shaderModule,
        entryPoint: "Vertex_Main" 
    };
    const badEntryPointPipelineDescsriptor = {
        vertexStage: badEntryPointDescriptor, 
        fragmentStage: fragmentStageDescriptor,
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(device, badEntryPointPipelineDescsriptor, "vertex shader entry point");
}, "Bad pipelines were not created");
</script>
</html>
